// apps/web/src/api/generated/schemas/notificationRead.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { NotificationReadCreatedAt } from './notificationReadCreatedAt';
import type { NotificationReadUpdatedAt } from './notificationReadUpdatedAt';
import type { NotificationReadUserId } from './notificationReadUserId';
import type { NotificationReadTenantId } from './notificationReadTenantId';
import type { NotificationReadNotificationType } from './notificationReadNotificationType';
import type { NotificationReadIsRead } from './notificationReadIsRead';
import type { NotificationReadReadAt } from './notificationReadReadAt';
import type { NotificationReadData } from './notificationReadData';

/**
 * Schema for reading Notification - replaces to_dict()
 */
export interface NotificationRead {
  /** Creation timestamp */
  createdAt?: NotificationReadCreatedAt;
  /** Last update timestamp */
  updatedAt?: NotificationReadUpdatedAt;
  /** Unique identifier for the resource */
  id: string;
  userId?: NotificationReadUserId;
  tenantId?: NotificationReadTenantId;
  title: string;
  message: string;
  notificationType?: NotificationReadNotificationType;
  isRead?: NotificationReadIsRead;
  readAt?: NotificationReadReadAt;
  data?: NotificationReadData;
}